dkg: verify private key before p2p setup
